03-API/requirements.txt
#pip install -r requirements.txt

blinker==1.6.2
click==8.1.3
Flask==2.3.2
Flask-Cors==3.0.10
future==0.18.3
importlib-metadata==6.6.0
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.24.3
pyserial==3.5
python-dotenv==1.0.0
PyYAML==6.0
scipy==1.10.1
serial==0.0.97
six==1.16.0
Werkzeug==2.3.6
zipp==3.15.0
